Allow custom log formats via a flag
This adds preliminary support for specifying a custom log format via
a command line flag.

Backs out the structured parsing of fields that are not always present
in logs to just the bare minimum that we can guarantee will be there.

Still need to remove Severity in lieu of Level.